+++ lib/cv2r.js
@@ -62,7 +62,7 @@ module.exports = async function create(opts = {}) {
 		await fs.writeFile(opts.output.spoiler,
 			global.spoiler.map(s => `"${s[0]}","${s[1]}","${s[2]}"`).join('\n'));
 		if (opts.rom) {
-			await fs.writeFile(output.rom, pm.patch(await fs.readFile(opts.rom)));
+			await fs.writeFile(output.rom, pm.generateRom(await fs.readFile(opts.rom)));
 		}
 	}
