+++ lib/index.js
@@ -13,5 +13,6 @@ module.exports = {
 	dir: {
 		patch: path.join(__dirname, '..', 'patch', 'optional'),
 		palette: path.join(__dirname, '..', 'patch', 'palette')
-	}
+	},
+	constants: require('../config/constants')
 };
