+++ lib/patch-manager.js
@@ -94,11 +94,16 @@ class PatchManager {
 		return patches;
 	}
 
+	getPatchMethods() {
+		const files = fs.readdirSync(path.join(__dirname, 'patch-types'));
+		return files.map(f => f.replace(/\.js$/, ''));
+	}
+
 	export() {
 		return order.reduce((a,c) => a.concat(this[c]), []);
 	}
 
-	patch(romBuffer) {
+	generateRom(romBuffer) {
 		order.forEach(o => {
 			this[o].forEach(entry => {
 				Object.keys(entry).forEach(offset => {
@@ -107,6 +112,44 @@ class PatchManager {
 			});
 		});
 		return romBuffer;
+	}
+
+	async generatePatch(name, files, opts = {}) {
+		try {
+			let { description = '', id, method = 'diff', output, type = 'optional' } = opts;
+			if (!output) {
+				if (!order.includes(type)) {
+					throw new Error(`Invalid patch type "${type}", must be: ${order.join(', ')}`);
+				}
+				output = path.join(__dirname, '..', 'patch', type);
+			}
+	
+			// validate name and necessary options
+			if (!name) { throw new Error('must provide a name for the patch'); }
+			if (!id) { 
+				id = name.toLowerCase().replace(/\s+/g, '-').replace(/[^a-z0-9-]/g, ''); 
+			}
+			if (!/[a-z0-9-]+/.test(id)) {
+				throw new Error(`invalid id ${id}, must contain only lowercase letters, numbers, or dashes (-)`);
+			}
+
+			// make sure we have a valid patch creation method
+			const patchFuncs = this.getPatchMethods();
+			const patchFuncList = patchFuncs.join(', ');
+			if (!method) { throw new Error(`must provide --m,method for patching, options include: ${patchFuncList}`); }
+			if (!patchFuncs.includes(method)) {
+				throw new Error(`invalid method ${method}, must be: ${patchFuncList}`);
+			}
+	
+			// get patch fata amd compose all info gathered into a patch in js module format
+			const pfOpts = { description, id, name, output };
+			const patchFunc = require(`./patch-types/${method}`);
+			await patchFunc(name, files, pfOpts);
+	
+		} catch (err) {
+			console.error(err);
+			process.exit(1);
+		}
 	}
 }
